requirement_file.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
click==8.0.3
Django==4.0
djangorestframework==3.13.1
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
Werkzeug==2.0.2